fix struct + refactor
